+++ src/utils/codeowners-command.ts
@@ -1,5 +1,5 @@
 import { Repositories } from '../app-config/app-config';
-import { invoke } from '@tauri-apps/api';
+import { invoke } from '@tauri-apps/api/core';
 import { useQuery, useQueryClient } from '@tanstack/react-query';
 import { useCallback } from 'react';