+++ src/utils/file-codeownets.ts
@@ -1,6 +1,6 @@
 import { Repositories } from '@/app-config/app-config';
 import { useQuery, useQueryClient } from '@tanstack/react-query';
-import { invoke } from '@tauri-apps/api';
+import { invoke } from '@tauri-apps/api/core';
 import { useCallback } from 'react';
 
 function getBranchFileCodeownersQueryKey(