+++ src/utils/useCheckUpdates.ts
@@ -1,5 +1,5 @@
 import { useQuery, useQueryClient } from '@tanstack/react-query';
-import { checkUpdate } from '@tauri-apps/api/updater';
+import { check } from '@tauri-apps/plugin-updater';
 import { useCallback } from 'react';
 
 const appUpdateQueryKey = ['app-update'];
@@ -8,7 +8,7 @@ const UPDATE_CHECK_DURATION = 1_000 * 60 * 60 * 24; // every day in ms
 export function useAppCheckUpdate() {
   const result = useQuery({
     queryKey: appUpdateQueryKey,
-    queryFn: () => checkUpdate(),
+    queryFn: () => check(),
     refetchInterval: UPDATE_CHECK_DURATION,
     staleTime: UPDATE_CHECK_DURATION, // It will trigger update once view
     refetchOnWindowFocus: false,