+++ vite.config.ts
@@ -25,19 +25,19 @@ export default defineConfig({
   // to access the Tauri environment variables set by the CLI with information about the current target
   envPrefix: [
     'VITE_',
-    'TAURI_PLATFORM',
-    'TAURI_ARCH',
-    'TAURI_FAMILY',
-    'TAURI_PLATFORM_VERSION',
-    'TAURI_PLATFORM_TYPE',
-    'TAURI_DEBUG',
+    'TAURI_ENV_PLATFORM',
+    'TAURI_ENV_ARCH',
+    'TAURI_ENV_FAMILY',
+    'TAURI_ENV_PLATFORM_VERSION',
+    'TAURI_ENV_PLATFORM_TYPE',
+    'TAURI_ENV_DEBUG',
   ],
   build: {
     // Tauri uses Chromium on Windows and WebKit on macOS and Linux
-    target: process.env.TAURI_PLATFORM == 'windows' ? 'chrome105' : 'safari13',
+    target: process.env.TAURI_ENV_PLATFORM == 'windows' ? 'chrome105' : 'safari13',
     // don't minify for debug builds
-    minify: !process.env.TAURI_DEBUG ? 'esbuild' : false,
+    minify: !process.env.TAURI_ENV_DEBUG ? 'esbuild' : false,
     // produce sourcemaps for debug builds
-    sourcemap: !!process.env.TAURI_DEBUG,
+    sourcemap: !!process.env.TAURI_ENV_DEBUG,
   },
 });